Add context API in OurProduct
