+++ data/db.json
@@ -67,16 +67,12 @@
       "rating": 5
     }
   ],
-  "cart": [
+  "cart": [],
+  "wishlist": [
     {
-      "id": "6",
-      "name": "Jr. Zoom Soccer Cleats",
-      "img": "product14.png"
-    },
-    {
-      "id": "1",
-      "name": "Breed Dry Dog Food",
-      "img": "product9.jfif"
+      "id": "2",
+      "name": "CANON EOS DSLR Camera",
+      "img": "product10.png"
     },
     {
       "id": "2",
@@ -87,13 +83,11 @@
       "id": "3",
       "name": "ASUS FHD Gaming Laptop",
       "img": "product11.png"
-    }
-  ],
-  "wishlist": [
+    },
     {
-      "id": "2",
-      "name": "CANON EOS DSLR Camera",
-      "img": "product10.png"
+      "id": "4",
+      "name": "Curology Product Set",
+      "img": "product12.png"
     }
   ],
   "fleshsale": [
