+++ src/components/Error/error-page.jsx
@@ -3,7 +3,7 @@ import { useNavigate, useRouteError } from "react-router-dom";
 export default function ErrorPage() {
   const error = useRouteError();
   const navigate = useNavigate();
-  console.log(error);
+  // console.log(error);
 
   return (
     <div id="error-page">
